zpaq: add a configuration header and get rid of build system cruft
The configuration header causes ZPAQ to only assume it's on Windows
if _WIN32 is defined.  This is necessary because OS X does not define
"unix", so it was attempting to include Windows-specific headers.

Also, it disables JIT on non-x86/amd64 CPUs, and enables it on all
x86/amd64 CPUs.  This replaces the old autotools macros which attempted
to detect whether SSE2 was really available which doesn't play well
when attempting to cross-compile.  In practice JIT should really be
enabled/disabled at runtime, but doing that efficiently would require
significant changes to ZPAQ and in practice we are unlikely to
encounter an x86/amd64 CPU which doesn't support SSE2.

https://github.com/quixdb/squash/issues/49

Signed-off-by: Evan Nemerson <evan@coeus-group.com>