Add thread_process API for plugins.
This makes it relatively easy to integrate libraries with APIs that
have a singe function call and expect you to implement readers and
writers which block.

The idea behind this API is that processing happens in a thread and
blocking is implemented with a call to a function (squash_stream_yield)
which yields control back to the main thread.  Once the user has had
a chance to update the input/output buffers control yields back to the
plugin's thread while the main thread blocks.  Potential users of this
API include ZPAQ, zling, brotli, CRUSH, etc.

Signed-off-by: Evan Nemerson <evan@nemerson.com>