Remoze LZMAT plugin.
LZMAT reads and writes outside of the buffer provided to it.  It does
not seem to be maintained—I sent an e-mail to the author a few days
ago and haven't heard back.  If anyone would like to fix the issue I
would be happy to re-add the plugin, but with LZHAM I don't think
LZMAT is particularly compelling anymore.

Signed-off-by: Evan Nemerson <evan@nemerson.com>