crush: fix an issue which was causing decoding errors on enwik8
When I adapted CRUSH for use as a library I removed the init_bits
function and moved the initialization into the context init function.
That function was called *inside* of a loop, so the two were not
equivalent.  This patch reintroduces init_bits and calls it in the
loop, just as the original does.

This also reintroduces C++ compatibility (mostly by casting malloc
results), and uses tabs instead of spaces more uniformly.

Signed-off-by: Evan Nemerson <evan@nemerson.com>