wflz: update to latest version, re-enable
wfLZ was hosted on Google Code, but has now moved to GitHub (and git
instead of subversion), so we can now drop the svn2github repo and use
the new official one.  The new version also contains some fixes to
avoid out-of-bounds accesses, which means we can re-enable the plugin
by default.

Signed-off-by: Evan Nemerson <evan@nemerson.com>