API/ABI BREAK: Use conformant array parameters for buffers
Conformant array parameters provide hints to the compiler and static
analyzers, potentially helping spot issues earlier.  For more
information on conformant array parameters, see
https://www.securecoding.cert.org/confluence/display/c/API05-C.+Use+conformant+array+parameters

Note that, since conformant array parameters require C99 or greater
(and do not work in C++), we use a macro to expand to the parameter
name only if we are using a C99 compiler, otherwise it expands to
nothing.

https://github.com/quixdb/squash/issues/89

Signed-off-by: Evan Nemerson <evan@nemerson.com>