Make sure submodules are available during configuration.
GitHub-generated archives do not include submodules, and therefore will
not work correctly.  GitHub is aware of the issue, their position is
that the current behavior is intentional.  There is no way to disable
these automatically generated archives.  In other words, an error
message which explains how to resolve the problem is about all I can do
right now.

https://github.com/quixdb/squash/issues/109

Signed-off-by: Evan Nemerson <evan@nemerson.com>