plugins: return SQUASH_BUFFER_FULL for undersized decompression buffers
This required improving the error handing in a lot of plugins, but in
the end *almost* all plugins can do it (which I found surprising).
The two which can't are:

 * lz4:lz4: https://github.com/Cyan4973/lz4/issues/156
 * ms-compress:lznt1: https://github.com/coderforlife/ms-compress/issues/17

I did have to modify my LZJB fork a bit, and LZO's behaviour is
disconcerting, but adding some checks to the bounds.c tests should
be comforting.

Signed-off-by: Evan Nemerson <evan@nemerson.com>