Preliminary support for OS X
The stable version of LZHAM doesn't work, but the devel repository
includes a fix, so I have switched Squash over to it.  If there are
reliability issues I will switch it back and simply disable LZHAM on
OS X.

This gets OS X to the point where it actually builds.  Unfortunately
it currently fails the unit tests (mostly threads with lznt1), which
will be difficult to figure out without access to an OS X system.