Fix compilation on older versions of clang
We want to add the -Wno-tautological-pointer-compare CFLAG when we are
compiling in debug mode so we can assert that values aren't NULL even
when we have a nonnull attribute that says so.  However, the flag isn't
available in all versions of clang that we want to support, and I can't
figure out when it was added, so we need to use the build system to
detect support.