Rewrite the buffered stream implementation.
The new code is much simpler, and performance is generally a bit
better.  The old version used a linked list of slices of memory, then
concatenated them all into a contiguous buffer (with malloc + memcpy)
once the input had been completely read.  This version just reallocs
a single slice, which saves a memcpy if the OS can mremap the buffer.