file: add the UTF-8-native squash_file_vprintf function back in
This lets us avoid the horrible (and slow) _vscwprintf hack on as long
as people aren't using wide-character strings.  Pretty much nobody uses
wide-character strings outside of Windows, so hitting there should be
fairly unlikely.  On Windows _vscwprintf exists natively and should be
fast enough, so it's not a major concern.