splice: use huge pages if available
Currently, this just checks to see if MAP_HUGETLB is defined and, if
so, uses it (and assumes 2 MiB pages).  This likely only works on
Linux, though I'm sure OS X and BSD have similar mechanisms.

This seems to fix the performance regression I have been seeing with
mapped files (I think it's due to a change in the kernel at around 4.1
or 4.2, but verifying that would be difficult).  2 MiB pages make a lot
of sense for Squash, though, so I'm not going to worry about it.

https://github.com/quixdb/squash/issues/125
https://github.com/quixdb/squash-benchmark/issues/17