options: make parsing integers and sizes safer
This should return SQUASH_RANGE if the value is outside of the the
allowable range instead of the more generic SQUASH_BAD_VALUE.

I do consider this to be an API break since the function's possible
return values are explicitly enumerated in the documentation.  However,
since Squash isn't API stable yet it is acceptable.