Use squash_malloc/free in several plugins.
There is still a lot of work to do here, though it's all pretty easy.
Density is somewhat interesting since there is no user data parameter,
but now that the first ones are done (and can be used as examples),
porting plugins for libraries which support custom allocators should be
pretty straightforward.