object: free memory in unref function
Previously, we required that subclasses provide a free function as a
callback to the base init function, then publicly expose a destroy
function.  This works, but it meant that every subclass had to provide
a free function which was always basically the same; call the destroy
function then free the associated memory.  This resulted in lots of
duplication in the plugins for very little gain.

Now we assume the memory was allocated with squash_malloc (or
squash_realloc, I suppose).  This removes the ability for a subclass
to use a non-standard allocator (which kind of defeats the purpose of
squash_malloc/realloc/free anyways), but means that they have to write
less code.