pithy: disable plugin due to security vulnerability
See https://github.com/nemequ/compfuzz/tree/master/libraries/pithy
for archives which trigger the issue(s).  When compiled without NDEBUG
there is an assertion failure, and when compiled with NDEBUG there is
an infinite loop.

The library author has not esponded to email about this issue, nor have
they responded to @jibsen's PR for Windows support, so I now consider
it to be abandoned.  I'll keep the plugin in-tree for a while in case
someone else is interested in taking over maintenance (of pithy, not
the plugin), but if not I'll likely remove the plugin some time before
1.0.