Add set_compiler_specific_flags CMake function, use it.
The main reason for this is that ICC and MSVC both accept similar style
arguments for disabling warnings, which will cause the compilers to be
confused about options intended for one another.  This function provides
a (relatively) easy way to limit attempts to add specific arguments to
only their intended compilers.