cli: make -LP list codecs as "plugin:codec"
This replaces the old format of the plugin name, a new line, then a
tab-indented list of codec(s).  The only real drawback to the new
format is it's slightly more difficult for humans to read, but since
the plugin:codec format is accepted as a codec name it makes it easy
to use bash to write a quick script to test things in every codec.