Add option to install unit tests.
Beyond being nice for certain distributions (autopkgtest for Debian
and Ubuntu, OpenEmbedded's ptest, Kyua for BSD, etc.), this provides
a convenient way for plugins not bundled with Squash to make use of
Squash's unit tests to check their plugins.

This is also a first step towards a separate repository for plugins
which don't meet Squash's requirements for inclusion in the main
repository.