Slice: When trying mmap, don't retry if mmaping was successful, and the codec failed
This prevents squash_splice from attempting to fall back to streaming if
the (de)compression fails: the fallback only occurs if mmapping fails.
In the worst case, this meant (de)compression would be attempted twice
for almost the entirety of the input, if the codec failed on data at the
end of the file.