Add ENABLE_SANITIZER cmake option, use it to enable asan on Travis
The old method of simply adding -fsanitize=* to C/CXXFLAGS wasn't
reliable.  It seemed to work in some places, but it hasn't worked on
Fedora for a while since it causes CMake's FindThreads module to not
bother pulling in pthreads.  With this, we can run FindThreads
*before* adding the -fsanitize=* flag.

This also enables AddressSanitizer on Travis, which actually works now
that density is no longer enabled.