Update µnit to latest version.
This version is a bit more robust when it comes to linking against
librt for clock_gettime.  If we can't find clock_gettime we don't
use it, even if the headers advertise it.  If we do manage to find
clock_gettime, we explicitly define a preprocessor symbol so the code
knows it is available.