Add Windows API plugin.
This is completely untested; it probably won't even compile.  However,
most of the work should be done, so if someone wants to pick it up and
finish it it should be pretty straightforward.  AFAIK the only thing
that really needs to be done is to fix the max_compressed_size
functions to better match whatever RtlCompressBuffer actually does.

See https://github.com/quixdb/squash/issues/54