squash.spec.in: Configure build as release with debug info
rpm tries to extract debug info from the built binary files but fails
with the current configuration because there is no debug info in the
binaries. Update the build type to be release with debug info so that
the debuginfo package which is generated has debuginfo and optimizations
are still performed.