plugin/lz4: Remove use of deprecated items
lz4 1.7.0 deprecated a handful of the compression functions and all of
the lz4f configuration enums. Update all usages to the new function
names or enums and add macros that convert the new name to the old name
when compiling against older versions. Remove
squash_lz4_compress_buffer_unsafe since all of the functions used were
deprecated and the replacements are the functions used by
squash_lz4_compress_buffer.