build: have cmake automatically download submodules if necessary
This is based on something I did for SIMDe, where there is no
autogen.sh.  There is less incentive to do this in Squash, but it
doesn't hurt, and it should make things nicer for the masochists who
prefer to invoke CMake directly (probably largely people on Windows
who lack a good shell for running autogen.sh/configure).